Corregir alarma que nunca se programaba: overflow de Int en medId
med.id se genera con Date.now() (13 dígitos), pero el módulo nativo lo
recibía como Int de 32 bits (máximo ~2.100 millones). La conversión
JS->Kotlin fallaba en silencio, por eso guardarYProgramar siempre caía en
el catch ("Se guardó con un aviso...") y la alarma nunca llegaba a
programarse. Se cambia medId a Long en todo el módulo (AlarmConstants,
AlarmScheduler, AlarmReceiver, AlarmRingService, ExpoMedAlarmModule).

+++ modules/expo-med-alarm/android/src/main/java/expo/modules/medalarm/AlarmReceiver.kt
@@ -12,8 +12,8 @@ import androidx.core.content.ContextCompat
  */
 class AlarmReceiver : BroadcastReceiver() {
   override fun onReceive(context: Context, intent: Intent) {
-    val medId = intent.getIntExtra(AlarmConstants.EXTRA_MED_ID, -1)
-    if (medId == -1) return
+    val medId = intent.getLongExtra(AlarmConstants.EXTRA_MED_ID, -1L)
+    if (medId == -1L) return
 
     val horaIndex = intent.getIntExtra(AlarmConstants.EXTRA_HORA_INDEX, 0)
     val intentos = intent.getIntExtra(AlarmConstants.EXTRA_INTENTOS, 0)
